xnu-7195.101.1
Imported from https://opensource.apple.com/tarballs/xnu/xnu-7195.101.1.tar.gz

+++ osfmk/kern/printf.c
@@ -365,7 +365,7 @@ __doprnt(
 
 		if (c == 'z' || c == 'Z') {
 			c = *++fmt;
-			if (sizeof(size_t) == sizeof(unsigned long)) {
+			if (sizeof(size_t) == sizeof(unsigned long long)) {
 				long_long = 1;
 			}
 		}
